refactor(hipsr): index-based constant lowering, co-located pattern, no opt-out
- Runtime: drop offset-based wrap_get_global + gpu_constants_blob_size;
  hipsr.constant lowers to hipdnn_ep_constant_get(ctx, index) again.
- Stamp index (walk-order ordinal into gpu_constants[]) on each constant in
  hipsr-externalize-constants, grouped with offset/size in the verifier.
- Move ConstantLowering into HipsrConstantOp.cpp so the op owns its LLVM
  lowering; HipsrDialectIR now links the LLVM conversion libraries.
- Mark the whole hipsr dialect illegal in the ConvertToLLVMPatternInterface.
- Remove the unused `externalize` opt-out flag (never set false in production);
  a consumer wanting a compile-time value replaces the op instead.
