+++ include/hip/Dialect/Hipsr/Transforms/Passes.td
@@ -46,9 +46,9 @@ def HipsrExternalizeConstantsPass
     Phase 2 of the hipsr constant subsystem. Runs in two stages:
 
     Phase 1 (always): walk every `hipsr.constant` in the module (across all
-    functions) that opts in (`shouldExternalize()` is true) and is not already
-    externalized; assign a 64-byte aligned cumulative offset and stamp the op
-    with `offset`/`size`.
+    functions) that is not already externalized; assign a 64-byte aligned
+    cumulative offset and stamp the op with `offset`/`size`/`index` (the
+    walk-order ordinal into the runtime constants array).
 
     Module-scoped by design: offsets accumulate into a single constants file shared by
     the whole module, so the pass must see every constant at once. A per-function
@@ -63,9 +63,8 @@ def HipsrExternalizeConstantsPass
     `writeConstantsBinToFileSystem`. When no FileSystem is injected (e.g.
     hip-mlir-opt) the pass is a pure IR transform (no file written).
 
-    Constants with `externalize = false` and ones already externalized are
-    skipped. This pass emits no module attributes; runtime metadata generation
-    is a separate follow-up.
+    Already-externalized constants are skipped (idempotent). This pass emits no
+    module attributes; runtime metadata generation is a separate follow-up.
   }];
   let options = [
     Option<"constantsFile", "constants-file", "std::string",
