+++ lib/Dialect/Hipsr/IR/CMakeLists.txt
@@ -32,4 +32,9 @@ target_link_libraries(HipsrDialectIR PUBLIC
   MLIRSupport
   MLIRArithDialect
   MLIRShapeDialect
+  # hipsr.constant owns its LLVM lowering (ConstantLowering in
+  # HipsrConstantOp.cpp), so the IR lib links the LLVM conversion machinery.
+  MLIRLLVMDialect
+  MLIRLLVMCommonConversion
+  MLIRTransforms
 )
