+++ lib/Dialect/Hipsr/IR/HipsrConstantOp.cpp
@@ -5,6 +5,14 @@
 
 #include "hip/Dialect/Hipsr/IR/HipsrConstantOp.h"
 
+#include "mlir/Conversion/LLVMCommon/MemRefBuilder.h"
+#include "mlir/Conversion/LLVMCommon/Pattern.h"
+#include "mlir/Conversion/LLVMCommon/TypeConverter.h"
+#include "mlir/Dialect/LLVMIR/FunctionCallUtils.h"
+#include "mlir/Dialect/LLVMIR/LLVMDialect.h"
+#include "mlir/Dialect/LLVMIR/LLVMTypes.h"
+#include "mlir/Transforms/DialectConversion.h"
+
 using namespace mlir;
 using namespace mlir::hipsr;
 
@@ -13,18 +21,21 @@ LogicalResult ConstantOp::verify() {
   bool hasSource = getSourceAttr() != nullptr;
   bool hasOffset = getOffsetAttr() != nullptr;
   bool hasSize = getSizeAttr() != nullptr;
+  bool hasIndex = getIndexAttr() != nullptr;
 
   // A data source is always present: exactly one of value or source (never
-  // both, never neither). Externalization only *adds* offset/size; it never
-  // removes the data source, so getDataValues() keeps working
+  // both, never neither). Externalization only *adds* offset/size/index; it
+  // never removes the data source, so getDataValues() keeps working
   // post-externalization.
-  //   {value} | {source} | {value, offset, size} | {source, offset, size}
-  if (hasValue == hasSource)
+  //   {value} | {source} | {value, offset, size, index} | {source, ...}
+  if (hasValue == hasSource) {
     return emitOpError("expected exactly one of {value} or {source}");
+  }
 
-  // offset/size are the externalized-location marker; set together or not.
-  if (hasOffset != hasSize)
-    return emitOpError("`offset` and `size` must be set together");
+  // offset/size/index are the externalized marker: all three set, or none.
+  if (hasOffset != hasSize || hasOffset != hasIndex) {
+    return emitOpError("`offset`, `size` and `index` must be set together");
+  }
 
   // The result type (ranked tensor or #hipsr.mem<device> memref) is enforced
   // by the Hipsr_TensorOrDeviceMemRef ODS constraint.
@@ -33,3 +44,117 @@ LogicalResult ConstantOp::verify() {
 
 #define GET_OP_CLASSES
 #include "hip/Dialect/Hipsr/IR/HipsrConstantOp.cpp.inc"
+
+namespace mlir {
+namespace hipsr {
+namespace {
+
+// Runtime symbol returning a constant's device pointer by its ordinal into the
+// gpu_constants[] array (hipdnn_ep_constant_get in hipdnn_ep_runtime_state.cpp;
+// same entry the hip.get_constant lowering targets).
+constexpr const char *kHipsrGetConstant = "hipdnn_ep_constant_get";
+
+// Lowers an externalized hipsr.constant to a @hipdnn_ep_constant_get(ctx,
+// index) runtime call whose returned device pointer is wrapped in a memref
+// descriptor. hipsr.constant carries no operands, so ctx is the enclosing
+// llvm.func's arg 0 (the runtime state pointer, per the EP ABI) and index is
+// the ordinal stamped by hipsr-externalize-constants.
+struct ConstantLowering : public ConvertOpToLLVMPattern<ConstantOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(ConstantOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    if (!op.isExternalized()) {
+      op.emitError(
+          "hipsr.constant reached LLVM lowering without externalization; "
+          "run -hipsr-externalize-constants first");
+      return failure();
+    }
+
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    MLIRContext *ctx = rewriter.getContext();
+    Type ptrType = LLVM::LLVMPointerType::get(ctx, 0);
+    Type i64Type = IntegerType::get(ctx, 64);
+
+    auto memRefType = dyn_cast<MemRefType>(op.getResult().getType());
+    if (!memRefType) {
+      op.emitError("hipsr.constant: externalized constant must have a memref "
+                   "result to lower to a runtime call");
+      return failure();
+    }
+
+    // ctx = the enclosing function's arg 0. Bail until the parent is llvm.func
+    // so the arg is already an !llvm.ptr (the conversion driver retries).
+    auto llvmFn = op->getParentOfType<LLVM::LLVMFuncOp>();
+    if (!llvmFn) {
+      return rewriter.notifyMatchFailure(op, "expected enclosing llvm.func");
+    }
+    if (llvmFn.getNumArguments() == 0) {
+      op.emitError("hipsr.constant: enclosing function has no arguments; "
+                   "expected the runtime context at arg 0");
+      return failure();
+    }
+    Value ctxArg = llvmFn.getArgument(0);
+
+    Value indexVal = LLVM::ConstantOp::create(
+        rewriter, loc, i64Type,
+        rewriter.getI64IntegerAttr(op.getIndexAttr().getInt()));
+
+    SmallVector<Type, 2> paramTypes = {ptrType, i64Type};
+    FailureOr<LLVM::LLVMFuncOp> funcOp = LLVM::lookupOrCreateFn(
+        rewriter, module, kHipsrGetConstant, paramTypes, ptrType);
+    if (failed(funcOp)) {
+      return failure();
+    }
+
+    SmallVector<Value, 2> args = {ctxArg, indexVal};
+    auto callOp = LLVM::CallOp::create(rewriter, loc, *funcOp, args);
+
+    // The runtime always returns a generic pointer (AS 0). Cast to the memref's
+    // address space (e.g. AS 1 = AMDGPU global memory) if needed.
+    FailureOr<unsigned> addrSpace =
+        getTypeConverter()->getMemRefAddressSpace(memRefType);
+    if (failed(addrSpace)) {
+      return failure();
+    }
+
+    Value dataPtr = callOp.getResult();
+    if (*addrSpace != 0) {
+      dataPtr = LLVM::AddrSpaceCastOp::create(
+          rewriter, loc, LLVM::LLVMPointerType::get(ctx, *addrSpace), dataPtr);
+    }
+
+    // Row-major sizes / strides for a static, identity-layout memref.
+    auto shape = memRefType.getShape();
+    SmallVector<Value, 4> sizes;
+    SmallVector<Value, 4> strides;
+    for (int64_t dim : shape) {
+      sizes.push_back(LLVM::ConstantOp::create(
+          rewriter, loc, i64Type, rewriter.getI64IntegerAttr(dim)));
+    }
+    int64_t stride = 1;
+    for (int i = shape.size() - 1; i >= 0; --i) {
+      strides.insert(strides.begin(), LLVM::ConstantOp::create(
+                                          rewriter, loc, i64Type,
+                                          rewriter.getI64IntegerAttr(stride)));
+      stride *= shape[i];
+    }
+
+    MemRefDescriptor desc = createMemRefDescriptor(
+        loc, memRefType, dataPtr, dataPtr, sizes, strides, rewriter);
+    rewriter.replaceOp(op, {desc});
+    return success();
+  }
+};
+
+} // namespace
+
+void populateHipsrConstantLoweringPatterns(const LLVMTypeConverter &converter,
+                                           RewritePatternSet &patterns) {
+  patterns.add<ConstantLowering>(converter);
+}
+
+} // namespace hipsr
+} // namespace mlir
